+++ blocklets/ai-studio/api/src/store/templates.ts
@@ -8,7 +8,16 @@ import { getTemplateIdFromPath, yjsToTemplate } from './projects';
 
 export const nextTemplateId = () => `${dayjs().format('YYYYMMDDHHmmss')}-${nanoid(6)}`;
 
-export type Role = 'system' | 'user' | 'assistant' | 'call-prompt' | 'call-api' | 'call-function' | 'call-dataset';
+export type Role =
+  | 'system'
+  | 'user'
+  | 'assistant'
+  | 'tool'
+  | 'call-prompt'
+  | 'call-api'
+  | 'call-function'
+  | 'call-dataset'
+  | 'call-macro';
 
 export const roles: Role[] = [
   'system',
@@ -18,12 +27,14 @@ export const roles: Role[] = [
   'call-api',
   'call-function',
   'call-dataset',
+  'call-macro',
 ];
 
 export interface PromptMessage {
   id: string;
   role?: Extract<Role, 'system' | 'user' | 'assistant'>;
   content?: string;
+  name?: string;
   visibility?: 'hidden';
 }
 
@@ -33,10 +44,12 @@ export interface CallPromptMessage {
   content?: undefined;
   output: string;
   template?: { id: string; name?: string };
-  parameters?: {
-    [key: string]: string | undefined;
-  };
+  parameters?: { [key: string]: string | undefined };
   visibility?: 'hidden';
+}
+
+export interface CallMacroMessage extends Omit<CallPromptMessage, 'role'> {
+  role: 'call-macro';
 }
 
 export interface CallAPIMessage {
@@ -70,12 +83,30 @@ export interface CallDatasetMessage {
   visibility?: 'hidden';
 }
 
-export type CallMessage = CallPromptMessage | CallAPIMessage | CallFuncMessage | CallDatasetMessage;
+export interface ToolsMessage {
+  id: string;
+  function: {
+    name: string;
+    description: string;
+    parameters?: {
+      type: 'object';
+      properties: { [key: string]: { type: any; description: 'string' } };
+      required: string[];
+    };
+  };
+  extraInfo: CallPromptMessage | CallAPIMessage | CallFuncMessage;
+}
 
-export type EditorPromptMessage = PromptMessage | CallMessage;
+export type CallMessage = CallPromptMessage | CallAPIMessage | CallFuncMessage | CallDatasetMessage | CallMacroMessage;
+
+export type EditorPromptMessage = PromptMessage | CallMessage | CallMacroMessage;
 
 export function isPromptMessage(message: any): message is PromptMessage {
-  return ['system', 'user', 'assistant'].includes(message?.role);
+  return ['system', 'user', 'assistant', 'tool'].includes(message?.role);
+}
+
+export function isCallMacroMessage(message: any): message is CallMacroMessage {
+  return message?.role === 'call-macro';
 }
 
 export function isCallPromptMessage(message: any): message is CallPromptMessage {
@@ -128,6 +159,7 @@ export interface Template {
     error?: { message: string };
     createdBy: string;
   }[];
+  tools?: ToolsMessage[];
 }
 
 export type Parameter = StringParameter | NumberParameter | SelectParameter | LanguageParameter | HoroscopeParameter;