+++ blocklets/ai-studio/src/pages/project/prompt-state.tsx
@@ -11,10 +11,11 @@ import { RecoilState, atom, useRecoilState } from 'recoil';
 
 import Mustache from '../../../api/src/libs/mustache';
 import { TemplateYjs } from '../../../api/src/store/projects';
-import { CallDatasetMessage, CallPromptMessage, Role } from '../../../api/src/store/templates';
+import { CallDatasetMessage, CallMacroMessage, CallPromptMessage, Role } from '../../../api/src/store/templates';
 import {
   isCallAPIMessage,
   isCallDatasetMessage,
+  isCallMacroMessage,
   isCallPromptMessage,
   isPromptMessage,
   isTemplate,
@@ -130,7 +131,7 @@ export function usePromptState({
       const editorState = await text2EditorState(content, role);
       setState((v) => ({ ...v, editorState }));
     }
-  }, [prompt?.data.content, prompt?.data.role, readOnly]);
+  }, [key, prompt?.data.content, prompt?.data.role, readOnly]);
 
   return { state, prompt, setEditorState, deletePrompt };
 }
@@ -145,7 +146,7 @@ export function useParameterState({
   projectId: string;
   gitRef: string;
   templateId: string;
-  prompt: CallPromptMessage | CallDatasetMessage;
+  prompt: CallPromptMessage | CallDatasetMessage | CallMacroMessage;
   param: string;
 }) {
   const content = isCallDatasetMessage(prompt) ? prompt.parameters?.query : prompt.parameters?.[param];
@@ -210,6 +211,15 @@ export function useEditorPicker({
 
   const getOptions = useCallback(
     (index?: number) => [
+      new ComponentPickerOption(t('call.list.macro'), {
+        keywords: ['execute', 'prompt'],
+        onSelect: (editor) => {
+          const variable = `${randomVariableNamePrefix}${randomId(5)}`;
+          const id = randomId();
+          addPrompt({ id, role: 'call-macro', output: variable }, index || 0);
+          editor.dispatchCommand(INSERT_VARIABLE_COMMAND, { name: variable });
+        },
+      }),
       new ComponentPickerOption(t('call.list.prompt'), {
         keywords: ['execute', 'prompt'],
         onSelect: (editor) => {
@@ -247,7 +257,7 @@ export function useEditorPicker({
         },
       }),
     ],
-    [addPrompt]
+    [addPrompt, t]
   );
 
   return { getOptions };
@@ -302,7 +312,8 @@ export function parseDirectivesOfTemplate(
     ...Object.values(template.prompts ?? {})
       .flatMap(({ data }) => {
         if (isPromptMessage(data)) return data.content;
-        if (isCallPromptMessage(data) && data.parameters) return Object.values(data.parameters);
+        if ((isCallPromptMessage(data) || isCallMacroMessage(data)) && data.parameters)
+          return Object.values(data.parameters);
         if (isCallAPIMessage(data) && data.url) {
           if (data.body) {
             return [data.url, data.body];
@@ -332,8 +343,8 @@ export function parseDirectivesOfTemplate(
   }
 
   if (includeEmptyPromptVariables && template.prompts) {
-    Object.values(template.prompts).forEach(({ data }) => {
-      if (isCallPromptMessage(data) && data.parameters) {
+    Object.values(template.prompts ?? {}).forEach(({ data }) => {
+      if ((isCallPromptMessage(data) || isCallMacroMessage(data)) && data.parameters) {
         Object.entries(data.parameters).forEach(([key, value]) => {
           if (!value) {
             directives.push({ type: 'variable', name: key });
@@ -426,7 +437,7 @@ export function usePromptsState({
         for (const { data } of Object.values(template.prompts)) {
           if (isPromptMessage(data) && data.content) {
             data.content = renameVariableByMustache(data.content, rename);
-          } else if (isCallPromptMessage(data) && data.parameters) {
+          } else if ((isCallPromptMessage(data) || isCallMacroMessage(data)) && data.parameters) {
             for (const param of Object.keys(data.parameters)) {
               const val = data.parameters[param];
               if (typeof val === 'string') {
@@ -453,6 +464,52 @@ export function usePromptsState({
   );
 
   return { addPrompt, renameVariable };
+}
+
+export function useToolsState({
+  projectId,
+  gitRef,
+  templateId,
+}: {
+  projectId: string;
+  gitRef: string;
+  templateId: string;
+}) {
+  const { store } = useStore(projectId, gitRef);
+
+  const file = store.files[templateId];
+  const template = isTemplate(file) ? file : undefined;
+
+  const addToolFunc = useCallback(
+    (func: NonNullable<TemplateYjs['tools']>[string]['data']) => {
+      if (!template) return;
+
+      const doc = (getYjsValue(template) as Map<any>).doc!;
+      doc.transact(() => {
+        template.tools ??= {};
+        template.tools[func.id] = { index: Object.keys(template.tools).length, data: func };
+
+        sortBy(Object.values(template.tools), (i) => i.index).forEach((i, index) => (i.index = index));
+      });
+    },
+    [template]
+  );
+
+  const deleteToolFunc = useCallback(
+    (id: string) => {
+      if (!template) return;
+
+      const doc = (getYjsValue(template) as Map<any>).doc!;
+
+      doc.transact(() => {
+        template.tools ??= {};
+        delete template.tools[id];
+      });
+    },
+    [template]
+  );
+
+  return { addToolFunc, deleteToolFunc };
 }
 
 export const randomId = customAlphabet('abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789');